% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbcl.R
\name{bootstrap_nbcl}
\alias{bootstrap_nbcl}
\title{Parametric bootstrap for NB-CL model}
\usage{
bootstrap_nbcl(
  object,
  B = 5000,
  correct_kappa = TRUE,
  seed = NULL,
  verbose = TRUE
)
}
\arguments{
\item{object}{An object of class "nbcl"}

\item{B}{Number of bootstrap samples (default 5000)}

\item{correct_kappa}{Whether to apply bias correction (default TRUE)}

\item{seed}{Random seed for reproducibility}

\item{verbose}{Print progress (default TRUE)}
}
\value{
An object of class "nbcl_boot" containing bootstrap results
}
\description{
Parametric bootstrap for NB-CL model
}
