% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCR.R
\name{buhlmann_straub}
\alias{buhlmann_straub}
\title{Bühlmann-Straub credibility estimator}
\usage{
buhlmann_straub(x, w, mu, sigma_sq, tau_sq)
}
\arguments{
\item{x}{Vector of observations (claim rates)}

\item{w}{Vector of exposure weights}

\item{mu}{Collective mean}

\item{sigma_sq}{Within-risk variance per unit weight}

\item{tau_sq}{Between-risk variance}
}
\value{
A list with credibility estimates and weights
}
\description{
Bühlmann-Straub credibility estimator
}
