% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCR.R
\name{chain_ladder}
\alias{chain_ladder}
\title{Chain-Ladder reserving method}
\usage{
chain_ladder(triangle, F_cum = NULL)
}
\arguments{
\item{triangle}{A matrix representing the run-off triangle (cumulative claims)}

\item{F_cum}{Optional vector of cumulative development proportions}
}
\value{
An object of class "reserve_estimate" with reserve estimates
}
\description{
The Chain-Ladder method corresponds to the credibility limit Z -> 1
(full weight on individual experience, diffuse prior).
}
