% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbcl.R
\name{compare_models}
\alias{compare_models}
\title{Compare Poisson, ODP, and NB-CL models}
\usage{
compare_models(triangle)
}
\arguments{
\item{triangle}{A matrix representing the run-off triangle}
}
\value{
A data frame comparing model fits
}
\description{
Compare Poisson, ODP, and NB-CL models
}
