% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCR.R
\name{compare_reserves}
\alias{compare_reserves}
\title{Compare all reserving methods}
\usage{
compare_reserves(triangle, exposure = NULL, prior_elr = NULL)
}
\arguments{
\item{triangle}{A matrix representing the run-off triangle (cumulative claims)}

\item{exposure}{Vector of exposures (optional)}

\item{prior_elr}{Prior expected loss ratio for BF (optional)}
}
\value{
A data frame comparing all methods
}
\description{
Compare all reserving methods
}
