% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCR.R
\name{credibility_bayes_equivalence}
\alias{credibility_bayes_equivalence}
\title{Demonstrate credibility = posterior mean}
\usage{
credibility_bayes_equivalence(n_obs, exposure, mu, tau_sq)
}
\arguments{
\item{n_obs}{Observed count}

\item{exposure}{Exposure}

\item{mu}{Prior mean of Lambda}

\item{tau_sq}{Prior variance of Lambda}
}
\value{
A list comparing credibility and Bayesian estimates
}
\description{
Demonstrate credibility = posterior mean
}
