% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCR.R
\name{estimate_bs_params}
\alias{estimate_bs_params}
\title{Estimate Bühlmann-Straub variance components}
\usage{
estimate_bs_params(x, w, sigma_sq = NULL)
}
\arguments{
\item{x}{Vector of observations (claim rates)}

\item{w}{Vector of exposure weights}

\item{sigma_sq}{Optional fixed process variance}
}
\value{
A list with estimated parameters
}
\description{
Estimate Bühlmann-Straub variance components
}
