% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbcl.R
\name{fit_nbcl}
\alias{fit_nbcl}
\title{Fit Negative Binomial Chain-Ladder model}
\usage{
fit_nbcl(triangle, link = "log")
}
\arguments{
\item{triangle}{A matrix representing the run-off triangle (incremental claims)}

\item{link}{Link function (default "log")}
}
\value{
An object of class "nbcl" containing the fitted model and metadata
}
\description{
Fit Negative Binomial Chain-Ladder model
}
