% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbcl.R
\name{kappa_corrected}
\alias{kappa_corrected}
\title{Extract bias-corrected kappa}
\usage{
kappa_corrected(object)
}
\arguments{
\item{object}{An object of class "nbcl"}
}
\value{
The bias-corrected dispersion parameter
}
\description{
Extract bias-corrected kappa
}
