% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbcl.R
\name{long_to_triangle}
\alias{long_to_triangle}
\title{Convert long format back to triangle}
\usage{
long_to_triangle(df, I = NULL, J = NULL)
}
\arguments{
\item{df}{A data frame with columns AY, DY, A}

\item{I}{Number of accident years}

\item{J}{Number of development years}
}
\value{
A matrix representing the run-off triangle
}
\description{
Convert long format back to triangle
}
