% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbcl.R
\name{lr_test_overdispersion}
\alias{lr_test_overdispersion}
\title{Likelihood ratio test for overdispersion}
\usage{
lr_test_overdispersion(object)
}
\arguments{
\item{object}{An object of class "nbcl"}
}
\value{
A list with test statistic, df, and p-value
}
\description{
Likelihood ratio test for overdispersion
}
