% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbcl.R
\name{plot.nbcl_profile}
\alias{plot.nbcl_profile}
\title{Plot profile likelihood}
\usage{
\method{plot}{nbcl_profile}(x, ...)
}
\arguments{
\item{x}{An object of class "nbcl_profile"}

\item{...}{Additional arguments passed to plot}
}
\description{
Plot profile likelihood
}
