% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbcl.R
\name{plot_diagnostics}
\alias{plot_diagnostics}
\title{Diagnostic plots for NB-CL model}
\usage{
plot_diagnostics(object, which = 1:4)
}
\arguments{
\item{object}{An object of class "nbcl"}

\item{which}{Which plots to produce (1 = residuals vs fitted,
2 = by AY, 3 = by DY, 4 = profile likelihood)}
}
\description{
Diagnostic plots for NB-CL model
}
