% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCR.R
\name{poisson_gamma_posterior}
\alias{poisson_gamma_posterior}
\title{Posterior distribution under Poisson-Gamma model}
\usage{
poisson_gamma_posterior(n_obs, exposure, alpha, beta)
}
\arguments{
\item{n_obs}{Observed count}

\item{exposure}{Exposure}

\item{alpha}{Prior shape parameter}

\item{beta}{Prior rate parameter}
}
\value{
A list with posterior parameters and summary statistics
}
\description{
Posterior distribution under Poisson-Gamma model
}
