% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbcl.R
\name{predict_interval}
\alias{predict_interval}
\title{Extract prediction intervals from bootstrap}
\usage{
predict_interval(object, level = 0.95)
}
\arguments{
\item{object}{An object of class "nbcl_boot"}

\item{level}{Confidence level (default 0.95)}
}
\value{
A data frame with point estimate and prediction intervals
}
\description{
Extract prediction intervals from bootstrap
}
