% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbcl.R
\name{predict_nbcl}
\alias{predict_nbcl}
\title{Predict lower triangle}
\usage{
predict_nbcl(object, newdata = NULL)
}
\arguments{
\item{object}{An object of class "nbcl"}

\item{newdata}{Optional new data for prediction}
}
\value{
A data frame with predicted values for the lower triangle
}
\description{
Predict lower triangle
}
