% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbcl.R
\name{print.nbcl}
\alias{print.nbcl}
\title{Print method for nbcl objects}
\usage{
\method{print}{nbcl}(x, ...)
}
\arguments{
\item{x}{An object of class "nbcl"}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for nbcl objects
}
