% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbcl.R
\name{print.nbcl_boot}
\alias{print.nbcl_boot}
\title{Print method for nbcl_boot objects}
\usage{
\method{print}{nbcl_boot}(x, ...)
}
\arguments{
\item{x}{An object of class "nbcl_boot"}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for nbcl_boot objects
}
