% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbcl.R
\name{profile_kappa}
\alias{profile_kappa}
\title{Profile likelihood for kappa}
\usage{
profile_kappa(object, kappa_range = NULL, n_points = 50, level = 0.95)
}
\arguments{
\item{object}{An object of class "nbcl"}

\item{kappa_range}{Range of kappa values to evaluate (default: auto)}

\item{n_points}{Number of points in the grid (default 50)}

\item{level}{Confidence level for CI (default 0.95)}
}
\value{
An object of class "nbcl_profile" with profile likelihood results
}
\description{
Profile likelihood for kappa
}
