% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomPI.R
\name{rdirichlet_}
\alias{rdirichlet_}
\title{Sample from a Dirichlet distribution}
\usage{
rdirichlet_(n, alpha)
}
\arguments{
\item{n}{Number of samples.}

\item{alpha}{Numeric vector of concentration parameters.}
}
\value{
Matrix with \code{n} rows and \code{length(alpha)} columns.
}
\description{
Sample from a Dirichlet distribution
}
\keyword{internal}
