% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbcl.R
\name{reserve_nbcl}
\alias{reserve_nbcl}
\title{Calculate reserve estimates}
\usage{
reserve_nbcl(object)
}
\arguments{
\item{object}{An object of class "nbcl"}
}
\value{
A list with total reserve and reserves by accident year
}
\description{
Calculate reserve estimates
}
