% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbcl.R
\name{residuals.nbcl}
\alias{residuals.nbcl}
\title{Extract residuals from NB-CL model}
\usage{
\method{residuals}{nbcl}(object, type = "pearson", ...)
}
\arguments{
\item{object}{An object of class "nbcl"}

\item{type}{Type of residuals: "pearson", "deviance", or "response"}

\item{...}{Additional arguments (ignored)}
}
\value{
A numeric vector of residuals
}
\description{
Extract residuals from NB-CL model
}
