% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCR.R
\name{run_ucr_simulation}
\alias{run_ucr_simulation}
\title{Run UCR simulation study}
\usage{
run_ucr_simulation(
  n_reps = 500,
  tau_sq_grid = c(0, 1e-04, 5e-04, 0.001, 0.005, 0.01),
  I = 10,
  J = 10,
  mu = 0.1,
  exposure = 1000,
  pi = NULL,
  verbose = TRUE
)
}
\arguments{
\item{n_reps}{Number of replications}

\item{tau_sq_grid}{Grid of tau^2 values to test}

\item{I}{Number of accident years}

\item{J}{Number of development periods}

\item{mu}{Mean frailty}

\item{exposure}{Exposure level}

\item{pi}{Development pattern}

\item{verbose}{Print progress}
}
\value{
A data frame with simulation results
}
\description{
Run UCR simulation study
}
