% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbcl.R
\name{summary.nbcl}
\alias{summary.nbcl}
\title{Summary method for nbcl objects}
\usage{
\method{summary}{nbcl}(object, ...)
}
\arguments{
\item{object}{An object of class "nbcl"}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary method for nbcl objects
}
