% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCR.R
\name{three_source_credibility}
\alias{three_source_credibility}
\title{Three-source credibility estimator}
\usage{
three_source_credibility(
  x,
  w,
  mu_pool,
  mu_prior,
  sigma_sq,
  tau_sq,
  tau_prior_sq
)
}
\arguments{
\item{x}{Vector of individual observations/rates}

\item{w}{Vector of exposure weights}

\item{mu_pool}{Pooled (collective) mean}

\item{mu_prior}{External prior mean}

\item{sigma_sq}{Process variance per unit weight}

\item{tau_sq}{Between-risk variance}

\item{tau_prior_sq}{Prior variance}
}
\value{
A list with credibility estimates and weights
}
\description{
Three-source credibility estimator
}
