% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbcl.R
\name{triangle_to_long}
\alias{triangle_to_long}
\title{Negative Binomial Chain-Ladder Model}
\usage{
triangle_to_long(triangle)
}
\arguments{
\item{triangle}{A matrix representing the run-off triangle (incremental claims)}
}
\value{
A data frame with columns AY, DY, A
}
\description{
Functions for fitting and predicting with the NB-CL model
}
\author{
Robin Van Oirbeek
}
